api/requirements.txt
# Django and REST Framework
Django==6.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# CORS handling
django-cors-headers==4.3.1

# Filtering and search
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# File handling
Pillow==10.2.0

# Date utilities
python-dateutil==2.8.2

# Environment variables
python-decouple==3.8

# API documentation (optional)
drf-yasg==1.21.7

# Testing
pytest==7.4.4
pytest-django==4.7.0

# Code quality
flake8==7.0.0
black==24.1.1

# Background tasks (optional)
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# Production server
gunicorn==21.2.0

# Monitoring (optional)
sentry-sdk==1.40.0

